requirements.txt
# Editable installs. Order matters: aegis-shared first (everything depends on
# it), then aegis-agents (the orchestrator service depends on it), then the
# per-service packages.
-e ./services/shared
-e ./agents
-e ./services/ingest
-e ./services/vision
-e ./services/orchestrator
-e ./services/dispatch
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
CacheControl==0.14.4
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
docstring_parser==0.18.0
fastapi==0.136.1
firebase_admin==7.4.0
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-aiplatform==1.148.1
google-cloud-appengine-logging==1.9.0
google-cloud-audit-log==0.5.0
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-cloud-firestore==2.27.0
google-cloud-logging==3.15.0
google-cloud-pubsub==2.37.0
google-cloud-resource-manager==1.17.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.73.1
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.13
importlib_metadata==8.7.1
iniconfig==2.3.0
msgpack==1.1.2
opentelemetry-api==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
packaging==26.1
pluggy==1.6.0
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
requests==2.33.1
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
structlog==25.5.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
watchfiles==1.1.1
websockets==16.0
zipp==3.23.1
